move checkForRetryableError to util.go
